+++ R/axes.R
@@ -272,7 +272,7 @@ handleticks <- function(data, ylim) {
 
 findlabelstep <- function(start, end, layout_factor, exponent = 1) {
   for (i in c(1, 2, 3, 4, 5, 10, 15, 20, 25, 30, 40, 50, 60, 70, 80, 90)) {
-    if ((floor((end - start) / (i*exponent)) + 1) < round(MAXXLABELS*layout_factor)) {
+    if (ceiling((end - start) / (i*exponent)) < round(MAXXLABELS*layout_factor)) {
       return(i*exponent)
     }
   }
@@ -301,7 +301,19 @@ getlayoutfactor <- function(layout) {
   }
 }
 
-xlabels.ts <- function(xlim, layout) {
+xlabels.ts_decade <- function(xlim, layout_factor) {
+  startyear <- floor(xlim[1]/5)*5
+  endyear <- ceiling(xlim[2]/5)*5
+  ticks <- seq(from = startyear, to = endyear, by = 10)
+  keep <- restrictlabels(ticks, layout_factor, xlim[2] < endyear && xlim[2] - xlim[1] > 3) # Only keep every 3rd or whatever label
+  labels <- ticks[keep]
+  at <- labels
+  # drop any labels that are outside the x limits
+  keep <- at >= xlim[1] & at <= xlim[2]
+  return(list(at = at[keep], labels = labels[keep], ticks = ticks))
+}
+
+xlabels.ts_year <- function(xlim, layout) {
   layout_factor <- getlayoutfactor(layout)
   startyear <- floor(xlim[1])
   endyear <- ceiling(xlim[2])
@@ -311,8 +323,73 @@ xlabels.ts <- function(xlim, layout) {
   labels <- ticks[keep]
   at <- labels + 0.5
   # drop any labels that are outside the x limits
-  keep <- at >= xlim[1] & at <= xlim[2]
+  keep <- at > xlim[1] & at < xlim[2]
   return(list(at = at[keep], labels = labels[keep], ticks = ticks))
+}
+
+xlabels.ts_quarter <- function(xlim) {
+  startquarter <- floor(xlim[1]*4)/4
+  endquarter <- ceiling(xlim[2]*4)/4
+  ticks <- seq(from = startquarter, to = (endquarter-0.25), by = 0.25)
+
+  # convert the labels to quarter names
+  qtrs <- 1 + 4*(ticks - floor(ticks))
+  labels <- substr(month.abb[qtrs*3], 1, 1)
+  at <- ticks + 0.5 * 0.25
+
+  # add years
+  startyear <- floor(xlim[1])
+  endyear <- floor(xlim[2])
+  labels <- c(labels, paste0("\n", seq(from = startyear, to = endyear, by = 1)))
+
+  # Manually adjust the years to adjust for partial first and last years
+  year_at <- seq(from = startyear + 0.5, to = endyear + 0.5, by = 1)
+  year_at[1] <- (xlim[1] + min(startyear + 1, xlim[2])) / 2
+  year_at[length(year_at)] <- (xlim[2] + min(endyear, xlim[2])) / 2
+  at <- c(at, year_at)
+
+  # drop any labels that are outside the x limits
+  keep <- at > xlim[1] & at < xlim[2]
+  return(list(at = at[keep], labels = labels[keep], ticks = ticks))
+}
+
+xlabels.ts_month <- function(xlim) {
+  startmonth <- floor(xlim[1]*12)/12
+  endmonth <- ceiling(xlim[2]*12)/12
+  ticks <- seq(from = startmonth, to = (endmonth-1/12), by = 1/12)
+
+  # convert the labels to month letters
+  months <- seq(from = 1 + (startmonth-floor(startmonth)) * 12, length.out = length(ticks))
+  labels <- substr(month.abb[1 + (months - 1) %% 12], 1, 1)
+  at <- ticks + 0.5 * 1/12
+
+  # add years
+  startyear <- floor(xlim[1])
+  endyear <- floor(xlim[2])
+  labels <- c(labels, paste0("\n", seq(from = startyear, to = endyear, by = 1)))
+
+  # Manually adjust the years to adjust for partial first and last years
+  year_at <- seq(from = startyear + 0.5, to = endyear + 0.5, by = 1)
+  year_at[1] <- (xlim[1] + min(startyear + 1, xlim[2])) / 2
+  year_at[length(year_at)] <- (xlim[2] + min(endyear, xlim[2])) / 2
+  at <- c(at, year_at)
+
+  # drop any labels that are outside the x limits
+  keep <- at > xlim[1] & at < xlim[2]
+  return(list(at = at[keep], labels = labels[keep], ticks = ticks))
+}
+
+xlabels.ts <- function(xlim, layout) {
+  layout_factor <- getlayoutfactor(layout)
+  if (xlim[2] - xlim[1] >= 50*layout_factor) {
+    return(xlabels.ts_decade(xlim, layout_factor))
+  } else if (xlim[2] - xlim[1] >= 3) {
+    return(xlabels.ts_year(xlim,layout))
+  } else if (xlim[2] - xlim[1] >= 1) {
+    return(xlabels.ts_quarter(xlim))
+  } else {
+    return(xlabels.ts_month(xlim))
+  }
 }
 
 xlabels.categorical <- function(xlim, xvar, layout, showall) {
@@ -400,6 +477,24 @@ is.scatter <- function(x) {
 
 defaultxscale <- function(xvars, ists) {
   if (is.numeric(xvars) && ists) {
+    start <- min(xvars, na.rm = TRUE)
+    end <- max(xvars, na.rm = TRUE)
+    if (end - start >= 3) {
+      return(c(floor(start),ceiling(end)))
+    } else if (end - start >= 1) {
+      return(c(floor(start*4)/4,ceiling(end*4)/4))
+    } else if (end - start > 0) {
+      return(c(floor(start*12)/12,ceiling(end*12)/12))
+    } else {
+      # Singleton observation, use years, not obvious what to do here?
+      if (start %% 1 == 0) {
+        # Right on a year, give half a year either side?
+        return(start-0.5,end+0.5)
+      } else {
+        # Otherwise just give the whole year that the singleton appears in
+        return(c(floor(start),ceiling(end)))
+      }
+    }
     return( c(floor(min(xvars, na.rm = TRUE)), ceiling(max(xvars, na.rm = TRUE))) )
   } else if (is.scatter(xvars)) {
     scale <- defaultscale(max(xvars, na.rm =TRUE), min(xvars, na.rm =TRUE))