+++ NEWS.md
@@ -16,6 +16,10 @@
  This means aesthetics can be _expressions_, which are evaluated within the environment
  of the provided data, rather than  just variables in the data. This means you
  can do things like `order = desc(variable)`, or `y = my_variable^2`.
+ * x-axis now supports decades, quarter and month spaced x-ticks for suitable length
+ samples ([#259](https://github.com/angusmoore/arphit/pull/259)). These are 
+ automatically selected if the data is short/long enough to warrant. Automatic
+ x-limits will choose partial years if the axis is in quarter or month terms.
 
 ## Breaking changes